requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
pillow==10.1.0
python-slugify==8.0.1